replace ffldb
