+++ database/ffldb/db.go
@@ -17,15 +17,14 @@ import (
 	"github.com/btcsuite/btcd/btcutil"
 	"github.com/btcsuite/btcd/chaincfg/chainhash"
 	"github.com/btcsuite/btcd/database"
+	"github.com/btcsuite/btcd/database/engine"
+	"github.com/btcsuite/btcd/database/engine/leveldb"
+	"github.com/btcsuite/btcd/database/engine/pebbledb"
 	"github.com/btcsuite/btcd/database/internal/treap"
 	"github.com/btcsuite/btcd/wire"
-	"github.com/syndtr/goleveldb/leveldb"
-	"github.com/syndtr/goleveldb/leveldb/comparer"
+
+	ldb "github.com/syndtr/goleveldb/leveldb"
 	ldberrors "github.com/syndtr/goleveldb/leveldb/errors"
-	"github.com/syndtr/goleveldb/leveldb/filter"
-	"github.com/syndtr/goleveldb/leveldb/iterator"
-	"github.com/syndtr/goleveldb/leveldb/opt"
-	"github.com/syndtr/goleveldb/leveldb/util"
 )
 
 const (
@@ -145,13 +144,13 @@ func convertErr(desc string, ldbErr error) database.Error {
 		code = database.ErrCorruption
 
 	// Database open/create errors.
-	case ldbErr == leveldb.ErrClosed:
+	case ldbErr == ldb.ErrClosed:
 		code = database.ErrDbNotOpen
 
 	// Transaction errors.
-	case ldbErr == leveldb.ErrSnapshotReleased:
+	case ldbErr == ldb.ErrSnapshotReleased:
 		code = database.ErrTxClosed
-	case ldbErr == leveldb.ErrIterReleased:
+	case ldbErr == ldb.ErrIterReleased:
 		code = database.ErrTxClosed
 	}
 
@@ -171,9 +170,9 @@ func copySlice(slice []byte) []byte {
 // and nested buckets of a bucket and implements the database.Cursor interface.
 type cursor struct {
 	bucket      *bucket
-	dbIter      iterator.Iterator
-	pendingIter iterator.Iterator
-	currentIter iterator.Iterator
+	dbIter      engine.Iterator
+	pendingIter engine.Iterator
+	currentIter engine.Iterator
 }
 
 // Enforce cursor implements the database.Cursor interface.
@@ -491,10 +490,10 @@ func cursorFinalizer(c *cursor) {
 // NOTE: The caller is responsible for calling the cursorFinalizer function on
 // the returned cursor.
 func newCursor(b *bucket, bucketID []byte, cursorTyp cursorType) *cursor {
-	var dbIter, pendingIter iterator.Iterator
+	var dbIter, pendingIter engine.Iterator
 	switch cursorTyp {
 	case ctKeys:
-		keyRange := util.BytesPrefix(bucketID)
+		keyRange := engine.BytesPrefix(bucketID)
 		dbIter = b.tx.snapshot.NewIterator(keyRange)
 		pendingKeyIter := newLdbTreapIter(b.tx, keyRange)
 		pendingIter = pendingKeyIter
@@ -509,7 +508,7 @@ func newCursor(b *bucket, bucketID []byte, cursorTyp cursorType) *cursor {
 		prefix := make([]byte, len(bucketIndexPrefix)+4)
 		copy(prefix, bucketIndexPrefix)
 		copy(prefix[len(bucketIndexPrefix):], bucketID)
-		bucketRange := util.BytesPrefix(prefix)
+		bucketRange := engine.BytesPrefix(prefix)
 
 		dbIter = b.tx.snapshot.NewIterator(bucketRange)
 		pendingBucketIter := newLdbTreapIter(b.tx, bucketRange)
@@ -523,26 +522,26 @@ func newCursor(b *bucket, bucketID []byte, cursorTyp cursorType) *cursor {
 		prefix := make([]byte, len(bucketIndexPrefix)+4)
 		copy(prefix, bucketIndexPrefix)
 		copy(prefix[len(bucketIndexPrefix):], bucketID)
-		bucketRange := util.BytesPrefix(prefix)
-		keyRange := util.BytesPrefix(bucketID)
+		bucketRange := engine.BytesPrefix(prefix)
+		keyRange := engine.BytesPrefix(bucketID)
 
 		// Since both keys and buckets are needed from the database,
 		// create an individual iterator for each prefix and then create
 		// a merged iterator from them.
 		dbKeyIter := b.tx.snapshot.NewIterator(keyRange)
 		dbBucketIter := b.tx.snapshot.NewIterator(bucketRange)
-		iters := []iterator.Iterator{dbKeyIter, dbBucketIter}
-		dbIter = iterator.NewMergedIterator(iters,
-			comparer.DefaultComparer, true)
+		iters := []engine.Iterator{dbKeyIter, dbBucketIter}
+		dbIter = engine.NewMergedIterator(iters,
+			engine.DefaultComparer, true)
 
 		// Since both keys and buckets are needed from the pending keys,
 		// create an individual iterator for each prefix and then create
 		// a merged iterator from them.
 		pendingKeyIter := newLdbTreapIter(b.tx, keyRange)
 		pendingBucketIter := newLdbTreapIter(b.tx, bucketRange)
-		iters = []iterator.Iterator{pendingKeyIter, pendingBucketIter}
-		pendingIter = iterator.NewMergedIterator(iters,
-			comparer.DefaultComparer, true)
+		iters = []engine.Iterator{pendingKeyIter, pendingBucketIter}
+		pendingIter = engine.NewMergedIterator(iters,
+			engine.DefaultComparer, true)
 	}
 
 	// Create the cursor using the iterators.
@@ -1860,6 +1859,8 @@ func (tx *transaction) Rollback() error {
 // the database.DB interface.  All database access is performed through
 // transactions which are obtained through the specific Namespace.
 type db struct {
+	dbType string
+
 	writeLock sync.Mutex   // Limit to one write transaction at a time.
 	closeLock sync.RWMutex // Make database close block while txns active.
 	closed    bool         // Is the database closed?
@@ -1875,7 +1876,7 @@ var _ database.DB = (*db)(nil)
 //
 // This function is part of the database.DB interface implementation.
 func (db *db) Type() string {
-	return dbType
+	return db.dbType
 }
 
 // begin is the implementation function for the Begin database method.  See its
@@ -2084,29 +2085,42 @@ func fileExists(name string) bool {
 	return true
 }
 
-// initDB creates the initial buckets and values used by the package.  This is
+// initDB creates the initial buckets and values used by the package. This is
 // mainly in a separate function for testing purposes.
-func initDB(ldb *leveldb.DB) error {
-	// The starting block file write cursor location is file num 0, offset
-	// 0.
-	batch := new(leveldb.Batch)
-	batch.Put(bucketizedKey(metadataBucketID, writeLocKeyName),
+func initDB(engine engine.Engine) error {
+	// The starting block file write cursor location is file num 0, offset 0.
+	tx, err := engine.NewTransaction()
+	if err != nil {
+		return err
+	}
+
+	// Insert the starting block file write cursor location.
+	err = tx.Put(bucketizedKey(metadataBucketID, writeLocKeyName),
 		serializeWriteRow(0, 0))
+	if err != nil {
+		return fmt.Errorf("failed to set writeLocKeyName: %w", err)
+	}
 
 	// Create block index bucket and set the current bucket id.
 	//
 	// NOTE: Since buckets are virtualized through the use of prefixes,
 	// there is no need to store the bucket index data for the metadata
-	// bucket in the database.  However, the first bucket ID to use does
+	// bucket in the database. However, the first bucket ID to use does
 	// need to account for it to ensure there are no key collisions.
-	batch.Put(bucketIndexKey(metadataBucketID, blockIdxBucketName),
+	err = tx.Put(bucketIndexKey(metadataBucketID, blockIdxBucketName),
 		blockIdxBucketID[:])
-	batch.Put(curBucketIDKeyName, blockIdxBucketID[:])
+	if err != nil {
+		return fmt.Errorf("failed to set blockIdxBucketName: %w", err)
+	}
 
-	// Write everything as a single batch.
-	if err := ldb.Write(batch, nil); err != nil {
-		str := fmt.Sprintf("failed to initialize metadata database: %v",
-			err)
+	err = tx.Put(curBucketIDKeyName, blockIdxBucketID[:])
+	if err != nil {
+		return fmt.Errorf("failed to set curBucketIDKeyName: %w", err)
+	}
+
+	// Apply the batch write.
+	if err := tx.Commit(); err != nil {
+		str := fmt.Sprintf("failed to initialize metadata database: %v", err)
 		return convertErr(str, err)
 	}
 
@@ -2115,7 +2129,7 @@ func initDB(ldb *leveldb.DB) error {
 
 // openDB opens the database at the provided path.  database.ErrDbDoesNotExist
 // is returned if the database doesn't exist and the create flag is not set.
-func openDB(dbPath string, network wire.BitcoinNet, create bool) (database.DB, error) {
+func openDB(dbType string, dbPath string, network wire.BitcoinNet, create bool) (database.DB, error) {
 	// Error if the database doesn't exist and the create flag is not set.
 	metadataDbPath := filepath.Join(dbPath, metadataDbName)
 	dbExists := fileExists(metadataDbPath)
@@ -2127,19 +2141,23 @@ func openDB(dbPath string, network wire.BitcoinNet, create bool) (database.DB, e
 	// Ensure the full path to the database exists.
 	if !dbExists {
 		// The error can be ignored here since the call to
-		// leveldb.OpenFile will fail if the directory couldn't be
+		// db Open will fail if the directory couldn't be
 		// created.
 		_ = os.MkdirAll(dbPath, 0700)
 	}
 
 	// Open the metadata database (will create it if needed).
-	opts := opt.Options{
-		ErrorIfExist: create,
-		Strict:       opt.DefaultStrict,
-		Compression:  opt.NoCompression,
-		Filter:       filter.NewBloomFilter(10),
+	var dbEngine engine.Engine
+	switch dbType {
+	case PebbleDB:
+		dbEngine = new(pebbledb.DB)
+	case LevelDB:
+		dbEngine = new(leveldb.DB)
+	default:
+		return nil, fmt.Errorf("driver %q is not registered", dbType)
 	}
-	ldb, err := leveldb.OpenFile(metadataDbPath, &opts)
+
+	err := dbEngine.Init(create, metadataDbPath)
 	if err != nil {
 		return nil, convertErr(err.Error(), err)
 	}
@@ -2147,16 +2165,16 @@ func openDB(dbPath string, network wire.BitcoinNet, create bool) (database.DB, e
 	// Create the block store which includes scanning the existing flat
 	// block files to find what the current write cursor position is
 	// according to the data that is actually on disk.  Also create the
-	// database cache which wraps the underlying leveldb database to provide
+	// database cache which wraps the underlying DB database to provide
 	// write caching.
 	store, err := newBlockStore(dbPath, network)
 	if err != nil {
 		return nil, convertErr(err.Error(), err)
 	}
-	cache := newDbCache(ldb, store, defaultCacheSize, defaultFlushSecs)
-	pdb := &db{store: store, cache: cache}
+	cache := newDbCache(dbEngine, store, defaultCacheSize, defaultFlushSecs)
+	db := &db{dbType: dbType, store: store, cache: cache}
 
 	// Perform any reconciliation needed between the block and metadata as
 	// well as database initialization, if needed.
-	return reconcileDB(pdb, create)
+	return reconcileDB(db, create)
 }
