+++ database/ffldb/dbcache.go
@@ -10,10 +10,8 @@ import (
 	"sync"
 	"time"
 
+	"github.com/btcsuite/btcd/database/engine"
 	"github.com/btcsuite/btcd/database/internal/treap"
-	"github.com/syndtr/goleveldb/leveldb"
-	"github.com/syndtr/goleveldb/leveldb/iterator"
-	"github.com/syndtr/goleveldb/leveldb/util"
 )
 
 const (
@@ -40,18 +38,18 @@ const (
 )
 
 // ldbCacheIter wraps a treap iterator to provide the additional functionality
-// needed to satisfy the leveldb iterator.Iterator interface.
+// needed to satisfy the engine.Iterator interface.
 type ldbCacheIter struct {
 	*treap.Iterator
 }
 
-// Enforce ldbCacheIterator implements the leveldb iterator.Iterator interface.
-var _ iterator.Iterator = (*ldbCacheIter)(nil)
+// Enforce ldbCacheIterator implements the engine.Iterator interface.
+var _ engine.Iterator = (*ldbCacheIter)(nil)
 
 // Error is only provided to satisfy the iterator interface as there are no
 // errors for this memory-only structure.
 //
-// This is part of the leveldb iterator.Iterator interface implementation.
+// This is part of the engine.Iterator interface implementation.
 func (iter *ldbCacheIter) Error() error {
 	return nil
 }
@@ -59,20 +57,20 @@ func (iter *ldbCacheIter) Error() error {
 // SetReleaser is only provided to satisfy the iterator interface as there is no
 // need to override it.
 //
-// This is part of the leveldb iterator.Iterator interface implementation.
-func (iter *ldbCacheIter) SetReleaser(releaser util.Releaser) {
+// This is part of the engine.Iterator interface implementation.
+func (iter *ldbCacheIter) SetReleaser(releaser engine.Releaser) {
 }
 
 // Release is only provided to satisfy the iterator interface.
 //
-// This is part of the leveldb iterator.Iterator interface implementation.
+// This is part of the engine.Iterator interface implementation.
 func (iter *ldbCacheIter) Release() {
 }
 
 // newLdbCacheIter creates a new treap iterator for the given slice against the
 // pending keys for the passed cache snapshot and returns it wrapped in an
 // ldbCacheIter so it can be used as a leveldb iterator.
-func newLdbCacheIter(snap *dbCacheSnapshot, slice *util.Range) *ldbCacheIter {
+func newLdbCacheIter(snap *dbCacheSnapshot, slice *engine.Range) *ldbCacheIter {
 	iter := snap.pendingKeys.Iterator(slice.Start, slice.Limit)
 	return &ldbCacheIter{Iterator: iter}
 }
@@ -81,14 +79,14 @@ func newLdbCacheIter(snap *dbCacheSnapshot, slice *util.Range) *ldbCacheIter {
 // cache and underlying database.
 type dbCacheIterator struct {
 	cacheSnapshot *dbCacheSnapshot
-	dbIter        iterator.Iterator
-	cacheIter     iterator.Iterator
-	currentIter   iterator.Iterator
+	dbIter        engine.Iterator
+	cacheIter     engine.Iterator
+	currentIter   engine.Iterator
 	released      bool
 }
 
-// Enforce dbCacheIterator implements the leveldb iterator.Iterator interface.
-var _ iterator.Iterator = (*dbCacheIterator)(nil)
+// Enforce dbCacheIterator implements the engine.Iterator interface.
+var _ engine.Iterator = (*dbCacheIterator)(nil)
 
 // skipPendingUpdates skips any keys at the current database iterator position
 // that are being updated by the cache.  The forwards flag indicates the
@@ -157,7 +155,7 @@ func (iter *dbCacheIterator) chooseIterator(forwards bool) bool {
 // First positions the iterator at the first key/value pair and returns whether
 // or not the pair exists.
 //
-// This is part of the leveldb iterator.Iterator interface implementation.
+// This is part of the engine.Iterator interface implementation.
 func (iter *dbCacheIterator) First() bool {
 	// Seek to the first key in both the database and cache iterators and
 	// choose the iterator that is both valid and has the smaller key.
@@ -169,7 +167,7 @@ func (iter *dbCacheIterator) First() bool {
 // Last positions the iterator at the last key/value pair and returns whether or
 // not the pair exists.
 //
-// This is part of the leveldb iterator.Iterator interface implementation.
+// This is part of the engine.Iterator interface implementation.
 func (iter *dbCacheIterator) Last() bool {
 	// Seek to the last key in both the database and cache iterators and
 	// choose the iterator that is both valid and has the larger key.
@@ -181,7 +179,7 @@ func (iter *dbCacheIterator) Last() bool {
 // Next moves the iterator one key/value pair forward and returns whether or not
 // the pair exists.
 //
-// This is part of the leveldb iterator.Iterator interface implementation.
+// This is part of the engine.Iterator interface implementation.
 func (iter *dbCacheIterator) Next() bool {
 	// Nothing to return if cursor is exhausted.
 	if iter.currentIter == nil {
@@ -197,7 +195,7 @@ func (iter *dbCacheIterator) Next() bool {
 // Prev moves the iterator one key/value pair backward and returns whether or
 // not the pair exists.
 //
-// This is part of the leveldb iterator.Iterator interface implementation.
+// This is part of the engine.Iterator interface implementation.
 func (iter *dbCacheIterator) Prev() bool {
 	// Nothing to return if cursor is exhausted.
 	if iter.currentIter == nil {
@@ -213,7 +211,7 @@ func (iter *dbCacheIterator) Prev() bool {
 // Seek positions the iterator at the first key/value pair that is greater than
 // or equal to the passed seek key.  Returns false if no suitable key was found.
 //
-// This is part of the leveldb iterator.Iterator interface implementation.
+// This is part of the engine.Iterator interface implementation.
 func (iter *dbCacheIterator) Seek(key []byte) bool {
 	// Seek to the provided key in both the database and cache iterators
 	// then choose the iterator that is both valid and has the larger key.
@@ -225,14 +223,14 @@ func (iter *dbCacheIterator) Seek(key []byte) bool {
 // Valid indicates whether the iterator is positioned at a valid key/value pair.
 // It will be considered invalid when the iterator is newly created or exhausted.
 //
-// This is part of the leveldb iterator.Iterator interface implementation.
+// This is part of the engine.Iterator interface implementation.
 func (iter *dbCacheIterator) Valid() bool {
 	return iter.currentIter != nil
 }
 
 // Key returns the current key the iterator is pointing to.
 //
-// This is part of the leveldb iterator.Iterator interface implementation.
+// This is part of the engine.Iterator interface implementation.
 func (iter *dbCacheIterator) Key() []byte {
 	// Nothing to return if iterator is exhausted.
 	if iter.currentIter == nil {
@@ -244,7 +242,7 @@ func (iter *dbCacheIterator) Key() []byte {
 
 // Value returns the current value the iterator is pointing to.
 //
-// This is part of the leveldb iterator.Iterator interface implementation.
+// This is part of the engine.Iterator interface implementation.
 func (iter *dbCacheIterator) Value() []byte {
 	// Nothing to return if iterator is exhausted.
 	if iter.currentIter == nil {
@@ -257,14 +255,14 @@ func (iter *dbCacheIterator) Value() []byte {
 // SetReleaser is only provided to satisfy the iterator interface as there is no
 // need to override it.
 //
-// This is part of the leveldb iterator.Iterator interface implementation.
-func (iter *dbCacheIterator) SetReleaser(releaser util.Releaser) {
+// This is part of the engine.Iterator interface implementation.
+func (iter *dbCacheIterator) SetReleaser(releaser engine.Releaser) {
 }
 
 // Release releases the iterator by removing the underlying treap iterator from
 // the list of active iterators against the pending keys treap.
 //
-// This is part of the leveldb iterator.Iterator interface implementation.
+// This is part of the engine.Iterator interface implementation.
 func (iter *dbCacheIterator) Release() {
 	if !iter.released {
 		iter.dbIter.Release()
@@ -277,7 +275,7 @@ func (iter *dbCacheIterator) Release() {
 // Error is only provided to satisfy the iterator interface as there are no
 // errors for this memory-only structure.
 //
-// This is part of the leveldb iterator.Iterator interface implementation.
+// This is part of the engine.Iterator interface implementation.
 func (iter *dbCacheIterator) Error() error {
 	return nil
 }
@@ -285,7 +283,7 @@ func (iter *dbCacheIterator) Error() error {
 // dbCacheSnapshot defines a snapshot of the database cache and underlying
 // database at a particular point in time.
 type dbCacheSnapshot struct {
-	dbSnapshot    *leveldb.Snapshot
+	dbSnapshot    engine.Snapshot
 	pendingKeys   *treap.Immutable
 	pendingRemove *treap.Immutable
 }
@@ -301,7 +299,7 @@ func (snap *dbCacheSnapshot) Has(key []byte) bool {
 	}
 
 	// Consult the database.
-	hasKey, _ := snap.dbSnapshot.Has(key, nil)
+	hasKey, _ := snap.dbSnapshot.Has(key)
 	return hasKey
 }
 
@@ -317,7 +315,7 @@ func (snap *dbCacheSnapshot) Get(key []byte) []byte {
 	}
 
 	// Consult the database.
-	value, err := snap.dbSnapshot.Get(key, nil)
+	value, err := snap.dbSnapshot.Get(key)
 	if err != nil {
 		return nil
 	}
@@ -338,9 +336,13 @@ func (snap *dbCacheSnapshot) Release() {
 // The slice parameter allows the iterator to be limited to a range of keys.
 // The start key is inclusive and the limit key is exclusive.  Either or both
 // can be nil if the functionality is not desired.
-func (snap *dbCacheSnapshot) NewIterator(slice *util.Range) *dbCacheIterator {
+func (snap *dbCacheSnapshot) NewIterator(slice *engine.Range) *dbCacheIterator {
+	dbIter, err := snap.dbSnapshot.NewIterator(slice.Start, slice.Limit)
+	if err != nil {
+		return nil
+	}
 	return &dbCacheIterator{
-		dbIter:        snap.dbSnapshot.NewIterator(slice, nil),
+		dbIter:        dbIter,
 		cacheIter:     newLdbCacheIter(snap, slice),
 		cacheSnapshot: snap,
 	}
@@ -354,8 +356,8 @@ func (snap *dbCacheSnapshot) NewIterator(slice *util.Range) *dbCacheIterator {
 // can commit transactions at will without incurring large performance hits due
 // to frequent disk syncs.
 type dbCache struct {
-	// ldb is the underlying leveldb DB for metadata.
-	ldb *leveldb.DB
+	// dbEngine is the underlying DB for metadata.
+	dbEngine engine.Engine
 
 	// store is used to sync blocks to flat files.
 	store *blockStore
@@ -396,7 +398,7 @@ type dbCache struct {
 //
 // The snapshot must be released after use by calling Release.
 func (c *dbCache) Snapshot() (*dbCacheSnapshot, error) {
-	dbSnapshot, err := c.ldb.GetSnapshot()
+	dbSnapshot, err := c.dbEngine.NewSnapshot()
 	if err != nil {
 		str := "failed to open transaction"
 		return nil, convertErr(str, err)
@@ -420,21 +422,21 @@ func (c *dbCache) Snapshot() (*dbCacheSnapshot, error) {
 // the transaction to be rolled back and are returned from this function.
 // Otherwise, the transaction is committed when the user-supplied function
 // returns a nil error.
-func (c *dbCache) updateDB(fn func(ldbTx *leveldb.Transaction) error) error {
+func (c *dbCache) updateDB(fn func(tx engine.Transaction) error) error {
 	// Start a leveldb transaction.
-	ldbTx, err := c.ldb.OpenTransaction()
+	tx, err := c.dbEngine.NewTransaction()
 	if err != nil {
 		return convertErr("failed to open ldb transaction", err)
 	}
 
-	if err := fn(ldbTx); err != nil {
-		ldbTx.Discard()
+	if err := fn(tx); err != nil {
+		tx.Discard()
 		return err
 	}
 
 	// Commit the leveldb transaction and convert any errors as needed.
-	if err := ldbTx.Commit(); err != nil {
-		return convertErr("failed to commit leveldb transaction", err)
+	if err := tx.Commit(); err != nil {
+		return convertErr("failed to commit db transaction", err)
 	}
 	return nil
 }
@@ -451,10 +453,10 @@ type TreapForEacher interface {
 // updates to the underlying database.
 func (c *dbCache) commitTreaps(pendingKeys, pendingRemove TreapForEacher) error {
 	// Perform all leveldb updates using an atomic transaction.
-	return c.updateDB(func(ldbTx *leveldb.Transaction) error {
+	return c.updateDB(func(tx engine.Transaction) error {
 		var innerErr error
 		pendingKeys.ForEach(func(k, v []byte) bool {
-			if dbErr := ldbTx.Put(k, v, nil); dbErr != nil {
+			if dbErr := tx.Put(k, v); dbErr != nil {
 				str := fmt.Sprintf("failed to put key %q to "+
 					"ldb transaction", k)
 				innerErr = convertErr(str, dbErr)
@@ -467,7 +469,7 @@ func (c *dbCache) commitTreaps(pendingKeys, pendingRemove TreapForEacher) error 
 		}
 
 		pendingRemove.ForEach(func(k, v []byte) bool {
-			if dbErr := ldbTx.Delete(k, nil); dbErr != nil {
+			if dbErr := tx.Delete(k); dbErr != nil {
 				str := fmt.Sprintf("failed to delete "+
 					"key %q from ldb transaction",
 					k)
@@ -630,12 +632,12 @@ func (c *dbCache) Close() error {
 		// Even if there is an error while flushing, attempt to close
 		// the underlying database.  The error is ignored since it would
 		// mask the flush error.
-		_ = c.ldb.Close()
+		_ = c.dbEngine.Close()
 		return err
 	}
 
 	// Close the underlying leveldb database.
-	if err := c.ldb.Close(); err != nil {
+	if err := c.dbEngine.Close(); err != nil {
 		str := "failed to close underlying leveldb database"
 		return convertErr(str, err)
 	}
@@ -647,9 +649,9 @@ func (c *dbCache) Close() error {
 // leveldb instance.  The cache will be flushed to leveldb when the max size
 // exceeds the provided value or it has been longer than the provided interval
 // since the last flush.
-func newDbCache(ldb *leveldb.DB, store *blockStore, maxSize uint64, flushIntervalSecs uint32) *dbCache {
+func newDbCache(dbEngine engine.Engine, store *blockStore, maxSize uint64, flushIntervalSecs uint32) *dbCache {
 	return &dbCache{
-		ldb:           ldb,
+		dbEngine:      dbEngine,
 		store:         store,
 		maxSize:       maxSize,
 		flushInterval: time.Second * time.Duration(flushIntervalSecs),
