+++ database/ffldb/ldbtreapiter.go
@@ -5,26 +5,25 @@
 package ffldb
 
 import (
+	"github.com/btcsuite/btcd/database/engine"
 	"github.com/btcsuite/btcd/database/internal/treap"
-	"github.com/syndtr/goleveldb/leveldb/iterator"
-	"github.com/syndtr/goleveldb/leveldb/util"
 )
 
 // ldbTreapIter wraps a treap iterator to provide the additional functionality
-// needed to satisfy the leveldb iterator.Iterator interface.
+// needed to satisfy the engine.Iterator interface.
 type ldbTreapIter struct {
 	*treap.Iterator
 	tx       *transaction
 	released bool
 }
 
-// Enforce ldbTreapIter implements the leveldb iterator.Iterator interface.
-var _ iterator.Iterator = (*ldbTreapIter)(nil)
+// Enforce ldbTreapIter implements the engine.Iterator interface.
+var _ engine.Iterator = (*ldbTreapIter)(nil)
 
 // Error is only provided to satisfy the iterator interface as there are no
 // errors for this memory-only structure.
 //
-// This is part of the leveldb iterator.Iterator interface implementation.
+// This is part of the engine.Iterator interface implementation.
 func (iter *ldbTreapIter) Error() error {
 	return nil
 }
@@ -32,14 +31,14 @@ func (iter *ldbTreapIter) Error() error {
 // SetReleaser is only provided to satisfy the iterator interface as there is no
 // need to override it.
 //
-// This is part of the leveldb iterator.Iterator interface implementation.
-func (iter *ldbTreapIter) SetReleaser(releaser util.Releaser) {
+// This is part of the engine.Iterator interface implementation.
+func (iter *ldbTreapIter) SetReleaser(releaser engine.Releaser) {
 }
 
 // Release releases the iterator by removing the underlying treap iterator from
 // the list of active iterators against the pending keys treap.
 //
-// This is part of the leveldb iterator.Iterator interface implementation.
+// This is part of the engine.Iterator interface implementation.
 func (iter *ldbTreapIter) Release() {
 	if !iter.released {
 		iter.tx.removeActiveIter(iter.Iterator)
@@ -51,7 +50,7 @@ func (iter *ldbTreapIter) Release() {
 // pending keys for the passed transaction and returns it wrapped in an
 // ldbTreapIter so it can be used as a leveldb iterator.  It also adds the new
 // iterator to the list of active iterators for the transaction.
-func newLdbTreapIter(tx *transaction, slice *util.Range) *ldbTreapIter {
+func newLdbTreapIter(tx *transaction, slice *engine.Range) *ldbTreapIter {
 	iter := tx.pendingKeys.Iterator(slice.Start, slice.Limit)
 	tx.addActiveIter(iter)
 	return &ldbTreapIter{Iterator: iter, tx: tx}
