+++ database/ffldb/reconcile.go
@@ -53,7 +53,7 @@ func reconcileDB(pdb *db, create bool) (database.DB, error) {
 	// Perform initial internal bucket and value creation during database
 	// creation.
 	if create {
-		if err := initDB(pdb.cache.ldb); err != nil {
+		if err := initDB(pdb.cache.dbEngine); err != nil {
 			return nil, err
 		}
 	}
