+++ database/ffldb/whitebox_test.go
@@ -179,7 +179,7 @@ func TestCornerCases(t *testing.T) {
 	// directory is needed.
 	testName := "openDB: fail due to file at target location"
 	wantErrCode := database.ErrDriverSpecific
-	idb, err := openDB(dbPath, blockDataNet, true)
+	idb, err := openDB(LevelDB, dbPath, blockDataNet, true)
 	if !checkDbError(t, testName, err, wantErrCode) {
 		if err == nil {
 			idb.Close()
@@ -191,7 +191,7 @@ func TestCornerCases(t *testing.T) {
 	// Remove the file and create the database to run tests against.  It
 	// should be successful this time.
 	_ = os.RemoveAll(dbPath)
-	idb, err = openDB(dbPath, blockDataNet, true)
+	idb, err = openDB(LevelDB, dbPath, blockDataNet, true)
 	if err != nil {
 		t.Errorf("openDB: unexpected error: %v", err)
 		return
@@ -215,14 +215,14 @@ func TestCornerCases(t *testing.T) {
 	_ = os.RemoveAll(filePath)
 
 	// Close the underlying leveldb database out from under the database.
-	ldb := idb.(*db).cache.ldb
-	ldb.Close()
+	db := idb.(*db).cache.dbEngine
+	db.Close()
 
 	// Ensure initialization errors in the underlying database work as
 	// expected.
 	testName = "initDB: reinitialization"
 	wantErrCode = database.ErrDbNotOpen
-	err = initDB(ldb)
+	err = initDB(db)
 	if !checkDbError(t, testName, err, wantErrCode) {
 		return
 	}
@@ -605,9 +605,9 @@ func TestFailureScenarios(t *testing.T) {
 	// Create a new database to run tests against.
 	dbPath := filepath.Join(os.TempDir(), "ffldb-failurescenarios")
 	_ = os.RemoveAll(dbPath)
-	idb, err := database.Create(dbType, dbPath, blockDataNet)
+	idb, err := database.Create(LevelDB, dbPath, blockDataNet)
 	if err != nil {
-		t.Errorf("Failed to create test database (%s) %v", dbType, err)
+		t.Errorf("Failed to create test database (%s) %v", LevelDB, err)
 		return
 	}
 	defer os.RemoveAll(dbPath)
